# docs/openapi/proximadb-openapi.yaml
components:
...
  schemas:
...
    CollectionV2Response:
      description: Collection details response
      properties:
        canonical_embedding_precision:
          description: |-
            Canonical embedding precision label for stored vectors.

            Stable string matching the proto `EmbeddingPrecision::as_str_name()`
            (e.g. "EMBEDDING_PRECISION_FP16"). `None` when the collection didn't
            set a non-default (Unspecified/Fp32) precision.
          type:
          - string
          - 'null'
        collection_id:
          description: Collection ID
          type: string
        created_at:
          description: Creation timestamp
          type: string
        dimension:
          description: Vector dimension
          format: int32
          minimum: 0
          type: integer
        distance_metric:
          description: Distance metric
          type: string
        engine:
          description: Storage engine
          type: string
        index_specs:
          description: |-
            Per-index config (HNSW/IVF params, is_primary) as persisted at create
            time. Mirrors the gRPC-v2 `GetCollection` `index_specs` (TD-122 parity).
          items:
            $ref: '#/components/schemas/IndexSpecOutput'
          type: array
        name:
          description: Collection name
          type: string
        proxima_record_enabled:
          description: Whether ProximaRecord is enabled
          type: boolean
        quantization:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/QuantizationConfigOutput'
            description: Quantization config as persisted at create time, or `None` when unset.
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/SchemaDefinition'
            description: Schema definition (if defined)
        stats:
          $ref: '#/components/schemas/CollectionStatsV2'
          description: Collection statistics
        updated_at:
          description: Last update timestamp
          type:
          - string
          - 'null'
      required:
      - collection_id
      - name
      - dimension
      - engine
      - distance_metric
      - proxima_record_enabled
      - stats
      - index_specs
      - created_at
      type: object
...
    CreateCollectionV2Request:
      description: |-
        Request to create a collection with schema support

        ## Example JSON

        ```json
        {
            "name": "products",
            "dimension": 768,
            "engine": "sst",
            "schema": {
                "columns": [
                    {"name": "category", "data_type": "text", "indexed": true},
                    {"name": "price", "data_type": "float", "filterable": true},
                    {"name": "description", "data_type": "text", "max_length": 10000}
                ],
                "enforcement": "hybrid",
                "allow_additional_fields": true
            },
            "enable_proxima_record": true
        }
        ```
      properties:
        canonical_embedding_precision:
          description: |-
            Canonical embedding precision for stored vectors

            Options: "fp32" (default), "fp16", "bf16", "int8", "uint8".
            Accepts the same string forms as the gRPC / Arrow Flight surfaces
            (e.g. "half", "float16", "EMBEDDING_PRECISION_FP16"). The DDL
            service applies the same normalisation, so SDKs and pgwire
            converge on the same enum discriminant.
          type:
          - string
          - 'null'
        dimension:
          description: Vector dimension (required)
          format: int32
          minimum: 1
          type: integer
        distance_metric:
          description: |-
            Distance metric for vector similarity

            Options: "cosine", "euclidean", "dot_product"
            Default: "cosine"
          type:
          - string
          - 'null'
        enable_proxima_record:
          description: |-
            Enable ProximaRecord support for this collection

            When enabled:
            - Records can use rich props and text_fields
            - Schema validation is applied at insert time
            - TEXT columns are stored in dedicated columnar format

            Default: false (backward compatible with v1)
          type:
          - boolean
          - 'null'
        engine:
          description: |-
            Storage engine selection

            Options: "auto", "sst", "helix", "viper", "swift", "nova", "raptor", "tst"
            Default: "auto" (system selects optimal engine)
          type:
          - string
          - 'null'
        index_configs:
          description: |-
            Index configurations (e.g. an explicit IVF or HNSW index).

            Restores v1 parity: the v1 proto create accepted `index_configs`, which
            drive `active_algorithm_for` (e.g. an IVF index → recall-tune dispatches
            to the IVF arm). When omitted, the engine selects a default (HNSW).
          items:
            $ref: '#/components/schemas/IndexConfigInput'
          type:
          - array
          - 'null'
        index_policy:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/IndexPolicyInput'
            description: |-
              Per-collection cold/warm search routing policy (ADR-028). When omitted,
              the system decides cost-derived (`mode = "auto"`). Owners with hard SLAs
              override here; query-time precedence is per-query `search_mode` > this
              policy > cost-model auto-default.
        initial_capacity:
          description: Initial capacity hint for pre-allocation
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
        name:
          description: Collection name (required)
          minLength: 1
          type: string
        quantization:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/QuantizationConfigInput'
            description: |-
              Quantization config (gRPC-v2 parity). When omitted, quantization is
              left unset (engine default).
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/SchemaDefinition'
            description: Schema definition with column types
        tags:
          description: |-
            Operator metadata tags, e.g. `"recall_target:0.95"`,
            `"target_vector_count:1000"`, `"modalities:text,image"`. Consumed by the
            recall advisor / route-health (`services/collection/recall_target.rs`).
          items:
            type: string
          type:
          - array
          - 'null'
      required:
      - name
      - dimension
      type: object
    CreateCollectionV2Response:
      description: Response for collection creation
      properties:
        collection_id:
          description: Collection ID (same as name)
          type: string
        created_at:
          description: Creation timestamp
          type: string
        dimension:
          description: Vector dimension
          format: int32
          minimum: 0
          type: integer
        engine:
          description: Selected storage engine
          type: string
        name:
          description: Collection name
          type: string
        proxima_record_enabled:
          description: Whether ProximaRecord is enabled
          type: boolean
        schema_id:
          description: Schema ID (if schema was defined)
          type:
          - string
          - 'null'
      required:
      - collection_id
      - name
      - dimension
      - engine
      - proxima_record_enabled
      - created_at
      type: object
...
    DeleteCollectionV2Response:
      description: Response for deleting a collection through the v2 API.
      properties:
        collection_id:
          description: Deleted collection ID.
          type: string
        success:
          description: Whether the delete request was accepted.
          type: boolean
      required:
      - success
      - collection_id
      type: object
...
    ErrorResponse:
      description: |-
        Canonical ProximaDB error envelope (`{ error: { type, message, code } }`).

        `request_id` (also returned in the `X-Request-ID` response header) is present
        whenever the request passed through the request-id middleware; quote it in
        bug reports.
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      required:
      - error
      type: object
...
    ListCollectionsV2Response:
      description: Response for listing collections
      properties:
        collections:
          description: List of collections
          items:
            $ref: '#/components/schemas/CollectionV2Summary'
          type: array
        has_more:
          description: Whether there are more results
          type: boolean
        limit:
          description: Limit used in this request
          format: int32
          minimum: 0
          type: integer
        offset:
          description: Offset used in this request
          format: int32
          minimum: 0
          type: integer
        total:
          description: Total count of collections
          format: int64
          minimum: 0
          type: integer
      required:
      - collections
      - total
      - limit
      - offset
      - has_more
      type: object
...
    QuantizationConfigInput:
      description: |-
        REST input for quantization config (mirrors proto `QuantizationConfig`;
        gRPC-v2 parity with `V2QuantizationConfig`).
      properties:
        enabled:
          description: Enable quantization for this collection.
          type:
          - boolean
          - 'null'
        strategy:
          description: 'Strategy: "smart_defaults" (default) | "minimal" | "aggressive" | "custom_levels".'
          type:
          - string
          - 'null'
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    email: vijay@anvaiops.com
    name: Vijaykumar Singh
  description: SDK-facing ProximaDB REST contract. This specification is v2-first and centers on ProximaRecord, typed schemas, record search, and collection lifecycle operations. Legacy v1 compatibility routes are intentionally not part of this publishable SDK surface.
  license:
    identifier: Apache-2.0
    name: Apache-2.0
  title: ProximaDB REST API
  version: 0.2.0
openapi: 3.1.0
paths:
...
  /api/v2/collections:
    get:
      description: |-
        List all collections with pagination.

        ## Query Parameters

        - `limit`: Maximum number of collections to return (default: 100)
        - `offset`: Offset for pagination (default: 0)
        - `include_stats`: Whether to include collection statistics

        ## Response

        Returns [`ListCollectionsV2Response`] with collection list.

        ## Errors

        - `500 Internal Server Error`: List operation failed
      operationId: listCollections
      parameters:
      - description: 'Maximum number of collections to return (default: 100)'
        in: query
        name: limit
        required: false
        schema:
          format: int32
          minimum: 0
          type: integer
      - description: 'Offset for pagination (default: 0)'
        in: query
        name: offset
        required: false
        schema:
          format: int32
          minimum: 0
          type: integer
      - description: Whether to include statistics
        in: query
        name: include_stats
        required: false
        schema:
          type: boolean
      - description: Optional explicit tenant selector. Applied only when there is no authenticated tenant context — a JWT tenant claim takes precedence, and a header that disagrees with the authenticated tenant is rejected. Absent ⇒ the default tenant. Tenant isolation is structural on the server; this header only selects the tenant.
        in: header
        name: X-Tenant-ID
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCollectionsV2Response'
          description: Collection page.
      summary: List collections.
      tags:
      - Collections
    post:
      description: |-
        Create a new collection with optional schema support.

        ## Request Body

        See [`CreateCollectionV2Request`] for the expected JSON format.

        ## Response

        Returns [`CreateCollectionV2Response`] with collection details.

        ## Errors

        - `400 Bad Request`: Invalid request or schema
        - `409 Conflict`: Collection already exists
        - `500 Internal Server Error`: Creation failed
      operationId: createCollection
      parameters:
      - description: Optional explicit tenant selector. Applied only when there is no authenticated tenant context — a JWT tenant claim takes precedence, and a header that disagrees with the authenticated tenant is rejected. Absent ⇒ the default tenant. Tenant isolation is structural on the server; this header only selects the tenant.
        in: header
        name: X-Tenant-ID
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionV2Request'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCollectionV2Response'
          description: Collection created.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
      summary: Create a collection with optional schema.
      tags:
      - Collections
  /api/v2/collections/{collection_id}:
    delete:
      description: |-
        Delete a collection by ID/name. This v2 route keeps SDK lifecycle methods on
        the ProximaRecord-era API while delegating to the existing collection control
        plane.
      operationId: deleteCollection
      parameters:
      - description: Collection name/ID.
        in: path
        name: collection_id
        required: true
        schema:
          type: string
      - description: Optional explicit tenant selector. Applied only when there is no authenticated tenant context — a JWT tenant claim takes precedence, and a header that disagrees with the authenticated tenant is rejected. Absent ⇒ the default tenant. Tenant isolation is structural on the server; this header only selects the tenant.
        in: header
        name: X-Tenant-ID
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteCollectionV2Response'
          description: Collection deleted.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      summary: Delete a collection.
      tags:
      - Collections
    get:
      description: |-
        Get collection details including schema and statistics.

        ## Path Parameters

        - `collection_id`: Collection name/ID

        ## Response

        Returns [`CollectionV2Response`] with collection details.

        ## Errors

        - `404 Not Found`: Collection does not exist
        - `500 Internal Server Error`: Retrieval failed
      operationId: getCollection
      parameters:
      - description: Collection name/ID.
        in: path
        name: collection_id
        required: true
        schema:
          type: string
      - description: Optional explicit tenant selector. Applied only when there is no authenticated tenant context — a JWT tenant claim takes precedence, and a header that disagrees with the authenticated tenant is rejected. Absent ⇒ the default tenant. Tenant isolation is structural on the server; this header only selects the tenant.
        in: header
        name: X-Tenant-ID
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionV2Response'
          description: Collection details.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      summary: Get collection details.
      tags:
      - Collections
...
  /api/v2/observability/namespaces/{namespace}/logs:
    parameters:
    - in: path
      name: namespace
      required: true
      schema:
        type: string
    post:
      operationId: ingestLog
      parameters:
      - description: Optional explicit tenant selector. Applied only when there is no authenticated tenant context — a JWT tenant claim takes precedence, and a header that disagrees with the authenticated tenant is rejected. Absent ⇒ the default tenant. Tenant isolation is structural on the server; this header only selects the tenant.
        in: header
        name: X-Tenant-ID
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: Ingested.
      summary: Ingest a log entry.
      tags:
      - Observability
...
security:
- bearerAuth: []
servers:
- description: Local ProximaDB server
  url: http://localhost:5678
...